requirements.txt
fastapi
uvicorn
aiohttp
pytubefix
pytest-asyncio
pytest
httpx
pytest-mock
pydantic
pydantic-settings
redis
fakeredis
asyncmock
instaloader
python-multipart
passlib[bcrypt]
bcrypt
PyJWT
python-jose
sqlalchemy[asyncio]
asyncpg
alembic
authlib
itsdangerous
aiosmtplib
pika
aio_pika
celery[redis]
prometheus_fastapi_instrumentator